0.1.1
This release backports b04405c506f3364e8bbe2b706220a9a6083a07bd, which fixes a native build error in recent versions of node.  v1.1.0 also merged this fix, additionally made the module context-aware and incremented the major version in order to stay in sync with the upstream repo.  However, there are dependents that have not updated their dependency version of this module, and are thus broken.  This release is intended to allow those dependents to build.

As semver-patch, this release does not include the context-aware change.

Dependents should update to v1 -- no further maintenance of this branch is planned.